Create rotation_number.cpp
Move the first few elements of an array to the end of the array, we call it the rotation of the array. Input a rotation of an ascending sorted array, and output the smallest element of the rotated array. For example, the array [3,4,5,1,2] is a rotation of [1,2,3,4,5], and the minimum value of the array is 1.  

Example 1:

Enter: [3,4,5,1,2]
Output: 1
Example 2:

Input: [2,2,2,0,1]
Output: 0

Source: LeetCode
Link: https://leetcode-cn.com/problems/xuan-zhuan-shu-zu-de-zui-xiao-shu-zi-lcof
The copyright belongs to the deduction network. Please contact the official authorization for commercial reprint, and please indicate the source for non-commercial reprint.