Create check_upper.py
Given a word, you need to determine whether the capitalization of the word is correct.

We define that the capitalization of words is correct in the following cases:

All letters are capitalized, such as "USA".
All letters in a word are not capitalized, such as "leetcode".
If the word contains more than one letter, only the first letter is capitalized, such as "Google".
Otherwise, we define that the word does not use capital letters correctly.

Example 1:

Enter: "USA"
Output: True
Example 2:

Enter: "FlaG"
Output: False
Note: The input is a non-empty word composed of uppercase and lowercase Latin letters.

Source: LeetCode
Link: https://leetcode-cn.com/problems/detect-capital
The copyright belongs to the deduction network. Please contact the official authorization for commercial reprint, and please indicate the source for non-commercial reprint.