Create route_string.py
Given two strings, A and B.

The rotation operation of A is to move the leftmost character of A to the rightmost. For example, if A ='abcde', the result after moving once is'bcdea'. If after several rotation operations, A can become B, then return to True.

Example 1:
Input: A ='abcde', B ='cdeab'
Output: true

Example 2:
Input: A ='abcde', B ='abced'
Output: false
note:

The length of A and B does not exceed 100.

Source: LeetCode
Link: https://leetcode-cn.com/problems/rotate-string
The copyright belongs to the deduction network. Please contact the official authorization for commercial reprint, and please indicate the source for non-commercial reprint.