Create max_time.py
Given a 4-digit copy, return the maximum time that can be set in accordance with the 24-hour clock.

The minimum 24-hour time is 00:00, and the maximum is 23:59. Starting from 00:00 (midnight), the longer it passes, the time interval.

The answer is returned as a string of length 5. If the validity time cannot be determined, an empty string is returned.



Example 1:

Enter: [1,2,3,4]
Output: "23:41"
Example 2:

Enter: [5,5,5,5]
Output: ""

Source: LeetCode
Link: https://leetcode-cn.com/problems/largest-time-for-given-digits
The copyright belongs to the deduction network. Please contact the official authorization for commercial reprint, and please indicate the source for non-commercial reprint.