Create max_sub.py
Enter an integer array with positive and negative numbers in the array. One or more consecutive integers in the array form a sub-array. Find the maximum value of the sum of all sub-arrays.

The required time complexity is O(n).
To
Example 1:

Input: nums = [-2,1,-3,4,-1,2,1,-5,4]
Output: 6
Explanation: The sum of consecutive sub-arrays [4,-1,2,1] is the largest, which is 6.
To
prompt:
1 <= arr.length <= 10^5
-100 <= arr[i] <= 100

Source: LeetCode
Link: https://leetcode-cn.com/problems/lian-xu-zi-shu-zu-de-zui-da-he-lcof
The copyright belongs to Lingkou Network. For commercial reprints, please contact the official authorization. For non-commercial reprints, please indicate the source.