Create minPathSum.cpp
LeetCode 64:
Given a m x n grid containing non-negative integers, please find a path from the upper left corner to the lower right corner, so that the sum of the numbers on the path is the smallest.

Note: You can only move one step down or right at a time.

Example:

enter:
[
[1,3,1],
   [1,5,1],
   [4,2,1]
]
Output: 7
Explanation: Because the total of the path 1→3→1→1→1 is the smallest.

Source: LeetCode
Link: https://leetcode-cn.com/problems/minimum-path-sum
The copyright belongs to Lingkou Network. For commercial reprints, please contact the official authorization. For non-commercial reprints, please indicate the source.