Create Straight.cpp
LeetCode 剑指 Offer 61. 扑克牌中的顺子:
Randomly draw 5 cards from the playing cards to determine if it is a straight, that is, if the 5 cards are consecutive. 2~10 are the numbers themselves, A is 1, J is 11, Q is 12, K is 13, and big and small are 0, which can be regarded as any number. A cannot think 14.



Example 1:

Input: [1,2,3,4,5]
Output: True


Example 2:

Input: [0,0,1,2,5]
Output: True


limit:

Length is 5

The number of spots takes the variable [0, 13].

Source: LeetCode
Link: https: //leetcode-cn.com/problems/bu-ke-pai-zhong-de-shun-zi-lcof
The copyright belongs to Lingkou Network. For commercial reprints, please contact the official authorization. For non-commercial reprints, please indicate the source.